order by uid instead of timestamp
findByTask must return the currently being worked on task in the db.
It did that by sorting by tstamp. This has the flaw of being imprecise.
If 2 tasks though a race condition are created at the same time,
than the order would be wrong (like the test demonstrates).

sorting by uid is much more reliable for that information.
It also prevents system clock discrepancy problems
since the uid is always in order.
And it also already has an index on it... which isn't used but hey.